move load and dump fbp to containers, make them batch_methods
